Add taskRunner function and import test and build_pr_array modules
Add run function and call taskRunner with input

Update main.ts to use taskRunner instead of runAsync and buildArray

Update imports in main.ts to import testtasklog and build_pr_array modules
