+++ src/task/test.ts
@@ -1,0 +1,5 @@
+const testtasklog = () => {
+    console.log("hello world");
+}
+
+export { testtasklog };
